refactor(nim-serving): reorganize into api/models and pages/deploymentWizard
Move NIM model utilities, types, and hooks from src/utils/ and
src/wizardFields/ into the api/models/ and pages/deploymentWizard/
patterns established by PR #7436. Remove section comments and
unnecessary re-exports. Keep NIMImageDependencies in the wizard field
component where it belongs, and NIMImagesData in the hook that
produces it.
