+++ packages/nim-serving/src/api/models/types.ts
@@ -1,0 +1,9 @@
+export type NIMModelInfo = {
+  name: string;
+  displayName?: string;
+  shortDescription?: string;
+  namespace?: string;
+  tags?: string[];
+  latestTag?: string;
+  updatedDate?: string;
+};
